requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
elasticsearch==8.12.0
boto3==1.34.0
faker==23.2.0
python-dotenv==1.0.0
pydantic==2.6.4
requests==2.31.0
